+++ ivmodels/tests/anderson_rubin.py
@@ -3,8 +3,7 @@ import scipy
 
 from ivmodels.models.kclass import KClass
 from ivmodels.quadric import Quadric
-from ivmodels.tests.utils import _check_test_inputs
-from ivmodels.utils import oproj, proj
+from ivmodels.utils import _check_inputs, oproj, proj
 
 
 def more_powerful_subvector_anderson_rubin_critical_value_function(
@@ -169,7 +168,7 @@ def anderson_rubin_test(
        guggenberger2012asymptotic
        guggenberger2019more
     """
-    Z, X, y, W, C, D, beta = _check_test_inputs(Z, X, y, W=W, C=C, D=D, beta=beta)
+    Z, X, y, W, C, D, beta = _check_inputs(Z, X, y, W=W, C=C, D=D, beta=beta)
 
     n, k = Z.shape
     mw, mc, md = W.shape[1], C.shape[1], D.shape[1]
@@ -295,7 +294,7 @@ def inverse_anderson_rubin_test(
     if not 0 < alpha < 1:
         raise ValueError("alpha must be in (0, 1).")
 
-    Z, X, y, W, C, D, _ = _check_test_inputs(Z, X, y, W=W, C=C, D=D)
+    Z, X, y, W, C, D, _ = _check_inputs(Z, X, y, W=W, C=C, D=D)
 
     n, k = Z.shape
     mx, mw, mc, md = X.shape[1], W.shape[1], C.shape[1], D.shape[1]
@@ -335,9 +334,6 @@ def inverse_anderson_rubin_test(
     A = S.T @ (S_proj - quantile * S_orth)
     b = -2 * (S_proj - quantile * S_orth).T @ y
     c = y.T @ (y_proj - quantile * y_orth)
-
-    if isinstance(c, np.ndarray):
-        c = c.item()
 
     coordinates = np.concatenate([np.arange(mx), np.arange(mx + mw, mx + mw + md)])
     return Quadric(A, b, c).project(coordinates)