+++ ivmodels/tests/wald.py
@@ -3,8 +3,7 @@ import scipy
 
 from ivmodels.models.kclass import KClass
 from ivmodels.quadric import Quadric
-from ivmodels.tests.utils import _check_test_inputs
-from ivmodels.utils import oproj, proj
+from ivmodels.utils import _check_inputs, oproj, proj
 
 
 def wald_test(
@@ -75,7 +74,7 @@ def wald_test(
         If the dimensions of the inputs are incorrect.
 
     """
-    Z, X, y, W, C, D, beta = _check_test_inputs(Z, X, y, W=W, C=C, D=D, beta=beta)
+    Z, X, y, W, C, D, beta = _check_inputs(Z, X, y, W=W, C=C, D=D, beta=beta)
 
     n, mx = X.shape
     mw, mc, md = W.shape[1], C.shape[1], D.shape[1]
@@ -178,7 +177,7 @@ def inverse_wald_test(
     if not 0 < alpha < 1:
         raise ValueError("alpha must be in (0, 1).")
 
-    Z, X, y, W, C, D, _ = _check_test_inputs(Z, X, y, W=W, C=C, D=D)
+    Z, X, y, W, C, D, _ = _check_inputs(Z, X, y, W=W, C=C, D=D)
 
     n, mx = X.shape
     mw, mc, md = W.shape[1], C.shape[1], D.shape[1]
@@ -216,8 +215,5 @@ def inverse_wald_test(
 
     b = -2 * A @ beta
     c = beta.T @ A @ beta - hat_sigma_sq * z_alpha
-
-    if isinstance(c, np.ndarray):
-        c = c.item()
 
     return Quadric(A, b, c)